web-ui: reg refactoring
