+++ meta-secret/web-cli/ui/src/views/VaultView.vue
@@ -23,17 +23,10 @@ onMounted(async () => {
     <p class="text-gray-400">Loading Vault Information...</p>
   </div>
 
-  <div v-else-if="jsAppState.isLocal || jsAppState.isVaultNotExists">
+  <div v-else-if="!jsAppState.isMember">
     <RegistrationComponent />
   </div>
   <div v-else-if="jsAppState.isMember">
     <VaultComponent />
-  </div>
-  <div v-else-if="jsAppState.isOutsider">
-    <div class="container mx-auto flex justify-center max-w-md pt-1 pb-4">
-      <div class="text-gray-900 dark:text-white">
-        <h1>Outsider!</h1>
-      </div>
-    </div>
   </div>
 </template>
